Total revamp of Anchorages, Trading Hubs, and Fortress Systems
Anchorages can now build more modules, with a preference for economic modules.

Trading Hubs try to position themselves over the Empire Capital, or as close to it as possible. It actively avoids locations where it could be cut off, or there is no trade.

Fortress Systems go for choke points, particularly choke points near empires that do not like them, and may be a threat. Further priority is given to points near hostile neighbours.